Finally fix incorrect flipping to quick settings.
Commit f6685c5 fixed this for the
two-finger pulldown. It reverted commit
c5e8dca in the process, which was
incorrect, but fixed another condition of this problem occuring
nevertheless. The actual problem the latter fix tried to fix was that
the flip decision was dependant on the timing of the view layout:
getMeasuredHeight() returned whatever was set in the last layout, which
might not have happened when hiding the view, so at the time of the next
ACTION_DOWN getMeasuredHeight() returned the height from the time the
panel was last visible.
Fix that problem by using the height scheduled for the next layout, so
we're always up-to-date.

Change-Id: I23b515a2dcf383a4b37569117a7ea47925ff02d9